+++ spec/unit/storage_spec.rb
@@ -1,4 +1,5 @@
 require File.dirname(__FILE__) + '/../spec_helper'
+require 'fileutils'
 
 describe AssetSync::Storage do
   include_context "mock Rails without_yml"
@@ -393,6 +394,46 @@ describe AssetSync::Storage do
       end
       storage.upload_file('assets/some_longer_path/local_image2.jpg')
     end
+
+    context 'config.gzip_compression is enabled' do
+      context 'when the file is a css file' do
+        it 'should upload the file' do
+          @config.gzip_compression = true
+
+          storage = AssetSync::Storage.new(@config)
+          allow(storage).to receive(:get_local_files).and_return(@local_files)
+          allow(storage).to receive(:get_remote_files).and_return(@remote_files)
+          # Pretend they all exist
+          allow(File).to receive(:open).and_return(file_like_object)
+          bucket = double
+          files = double
+          allow(storage).to receive(:bucket).and_return(bucket)
+          allow(bucket).to receive(:files).and_return(files)
+
+          expect(files).to receive(:create).with({ body: file_like_object, content_type: "text/css", key: "assets/local.css", public: true }).once
+          storage.upload_file('assets/local.css')
+        end
+      end
+
+      context 'when the file is a gz file' do
+        it 'should not upload the file' do
+          @config.gzip_compression = true
+
+          storage = AssetSync::Storage.new(@config)
+          allow(storage).to receive(:get_local_files).and_return(@local_files)
+          allow(storage).to receive(:get_remote_files).and_return(@remote_files)
+          # Pretend they all exist
+          allow(File).to receive(:open).and_return(file_like_object)
+          bucket = double
+          files = double
+          allow(storage).to receive(:bucket).and_return(bucket)
+          allow(bucket).to receive(:files).and_return(files)
+
+          expect(files).to_not receive(:create)
+          storage.upload_file('assets/local.css.gz')
+        end
+      end
+    end
   end
 
   describe '#delete_extra_remote_files' do
@@ -447,6 +488,54 @@ describe AssetSync::Storage do
         expect(file).to receive(:destroy)
 
         storage.delete_extra_remote_files
+      end
+    end
+  end
+
+  describe '#get_local_files' do
+    around(:each) do |example|
+      Dir.mktmpdir do |public_path|
+        @public_path = public_path
+        example.call
+      end
+    end
+
+    before(:each) do
+      @config = AssetSync::Config.new
+      @config.public_path = @public_path
+      @config.prefix = 'assets'
+      @storage = AssetSync::Storage.new(@config)
+
+      Dir.mkdir("#{@public_path}/assets")
+    end
+
+    context 'with empty directory' do
+      it 'has no files' do
+        expect(@storage.get_local_files).to eq([])
+      end
+    end
+
+    context 'with non-empty directory' do
+      before(:each) do
+        FileUtils.touch("#{@public_path}/assets/application.js")
+      end
+
+      it 'lists available files' do
+        expect(@storage.get_local_files).to eq([
+          'assets/application.js'
+        ])
+      end
+
+      context 'with trailing slash on asset prefix' do
+        before(:each) do
+          @config.prefix = 'assets/'
+        end
+
+        it 'lists available files with single slashes' do
+          expect(@storage.get_local_files).to eq([
+            'assets/application.js'
+          ])
+        end
       end
     end
   end